Make mock binaries independently reusable
