+++ cloudman/helmsman/tests/mock_helm.py
@@ -1,12 +1,15 @@
 import argparse
 import csv
 from io import StringIO
-from unittest.mock import patch
 import uuid
 import yaml
 
 
-class MockHelmParser(object):
+class MockHelm(object):
+    """
+    A mock version of the helm binary. Maintains an in-memory database
+    to simulate helm commands.
+    """
 
     def __init__(self):
         self.revision_history = [
@@ -38,21 +41,9 @@ class MockHelmParser(object):
                                           "CHART", "APP VERSION",
                                           "DESCRIPTION"]
         self.repo_list_field_names = ["NAME", "URL"]
+        self.parser = self._create_parser()
 
-    def can_parse(self, command):
-        if isinstance(command, list):
-            prog = command[0]
-            if prog.startswith("helm"):
-                return True
-        return False
-
-    @staticmethod
-    def extra_patches():
-        return [patch(
-            'helmsman.clients.helm_client.HelmClient._check_environment',
-            return_value=True)]
-
-    def parse_command(self, command):
+    def _create_parser(self):
         parser = argparse.ArgumentParser(prog='helm')
 
         subparsers = parser.add_subparsers(help='Available Commands')
@@ -142,8 +133,11 @@ class MockHelmParser(object):
         parser_list.add_argument('release', type=str, help='release name')
         parser_list.set_defaults(func=self._helm_delete)
 
+        return parser
+
+    def run_command(self, command):
         # evaluate command
-        args = parser.parse_args(command[1:])
+        args = self.parser.parse_args(command[1:])
         return args.func(args)
 
     def _helm_list(self, args):
@@ -271,117 +265,3 @@ class MockHelmParser(object):
             return 'Error: release: "%s" not found' % args.release
         self.chart_database.pop(args.release, None)
 
-
-class MockKubectlParser(object):
-
-    def __init__(self):
-        self.namespace_info = {
-            'NAME': 'default',
-            'STATUS': 'Active',
-            'AGE': '2d1h'
-        }
-        self.namespace_database = {'default': self.namespace_info}
-        self.namespace_list_field_names = ["NAME", "STATUS", "AGE"]
-
-    def can_parse(self, command):
-        if isinstance(command, list):
-            prog = command[0]
-            if prog.startswith("kubectl"):
-                if 'namespace' in command or 'namespaces' in command:
-                    return True
-        return False
-
-    @staticmethod
-    def extra_patches():
-        return [patch(
-          'helmsman.clients.k8s_client.KubeClient._check_environment',
-          return_value=True)]
-
-    def parse_command(self, command):
-        parser = argparse.ArgumentParser(prog='kubectl')
-        subparsers = parser.add_subparsers(help='Available Commands')
-
-        # kubectl get namespaces
-        parser_get = subparsers.add_parser('get',
-                                            help='list')
-        subparsers_get = parser_get.add_subparsers(help='Resources to get')
-        parser_list_ns = subparsers_get.add_parser('namespaces', help='List namespaces')
-        parser_list_ns.set_defaults(func=self._kubectl_get_namespaces)
-
-        # kubectl create namespace
-        parser_create = subparsers.add_parser('create',
-                                            help='create')
-        subparsers_create = parser_create.add_subparsers(help='Resources to create')
-        parser_create_ns = subparsers_create.add_parser('namespace',
-                                            help='create a namespace')
-        parser_create_ns.add_argument(
-            'namespace', type=str, help='namespace name')
-        parser_create_ns.set_defaults(func=self._kubectl_create_namespace)
-
-        # Kubectl delete namespace
-        parser_delete = subparsers.add_parser('delete',
-                                            help='delete')
-        subparsers_delete = parser_delete.add_subparsers(help='Resources to create')
-        parser_delete_ns = subparsers_delete.add_parser('namespace', help='delete a namespace')
-        parser_delete_ns.add_argument('namespace', type=str, help='namespace name')
-        parser_delete_ns.set_defaults(func=self._kubectl_delete_namespace)
-
-        # evaluate command
-        args = parser.parse_args(command[1:])
-        return args.func(args)
-
-    def _kubectl_get_namespaces(self, args):
-        # pretend to succeed
-        with StringIO() as output:
-            writer = csv.DictWriter(output,
-                                    fieldnames=self.namespace_list_field_names,
-                                    delimiter=" ", extrasaction='ignore')
-            writer.writeheader()
-            for release in self.namespace_database.values():
-                # Write data about the latest revision for each chart
-                writer.writerow(release)
-            return output.getvalue()
-
-    def _kubectl_create_namespace(self, args):
-        name = args.namespace
-        details = {
-            'NAME': name,
-            'STATUS': 'Active',
-            'AGE': '1d',
-        }
-        self.namespace_database[name] = details
-        return details
-
-    def _kubectl_delete_namespace(self, args):
-        name = args.namespace
-        details = self.namespace_database.get(name)
-        if not details:
-            return 'Error: namespace: "%s" not found' % name
-        self.namespace_database.pop(name, None)
-
-
-class MockClient(object):
-
-    """ Mocks all calls to the helm and kubectl commands"""
-    def __init__(self, testcase):
-        self.parsers = [MockHelmParser(), MockKubectlParser()]
-        self.extra_patches = []
-        for parser in self.parsers:
-            self.extra_patches += parser.extra_patches()
-        self.patch1 = patch('helmsman.clients.helpers.run_command',
-                            self.mock_run_command)
-        self.patch1.start()
-        testcase.addCleanup(self.patch1.stop)
-        for each in self.extra_patches:
-            each.start()
-            testcase.addCleanup(each.stop)
-
-    def mock_run_command(self, command, shell=False):
-        for parser in self.parsers:
-            if parser.can_parse(command):
-                return parser.parse_command(command)
-            # elif prog.startswith("kubectl create"):
-            #     # pretend to succeed
-            #     pass
-        else:
-            raise Exception("Unrecognised command: {0}".format(str(command)))
